requirements.txt
# requirements.txt
Django
djangorestframework
django-cors-headers
django-filter
pytest-django
